Moved node.py into baselib
