+++ openquake/calculators/tests/scenario_test.py
@@ -25,7 +25,7 @@ from nose.plugins.attrib import attr
 from openquake.qa_tests_data.scenario import (
     case_1, case_2, case_3, case_4, case_5, case_6, case_7, case_8, case_9)
 
-from openquake.commonlib.node import floatformat
+from openquake.baselib.node import floatformat
 from openquake.baselib.general import get_array
 from openquake.calculators.tests import CalculatorTestCase
