+++ openquake/commonlib/tests/writers_test.py
@@ -22,7 +22,7 @@ import tempfile
 from io import BytesIO
 from openquake.baselib.performance import memory_info
 from openquake.commonlib.writers import write_csv
-from openquake.commonlib.node import Node, tostring, StreamingXMLWriter
+from openquake.baselib.node import Node, tostring, StreamingXMLWriter
 from xml.etree import ElementTree as etree
 
 import numpy
