+++ openquake/engine/tools/correct_complex_sources.py
@@ -23,7 +23,7 @@ import sys
 import numpy
 
 from openquake.commonlib import nrml
-from openquake.commonlib.node import node_from_xml
+from openquake.baselib.node import node_from_xml
 
 from openquake.hazardlib.geo.line import Line
 from openquake.hazardlib.geo.point import Point
